Last comit for proper demonstration
Still very slopy, terrible lack of performance and other detail, but could be solve easely in multiple way. Just a lack of time to make it clean for production. Afterall, it's just a user-case study, and so there is not a urgent need about performance.
BUG FREE, more resilient. Still few exception not handle (by few, I mean... well, quite a lot)